+++ packages/dialtone-tokens/dialtone-transforms.js
@@ -4,9 +4,9 @@
  * because sd-transforms handles the transforms for CSS.
  */
 
-import Color from 'tinycolor2';
+import { colorModifiersFilter, DeviceObjectFormat, deviceTransformColorModifiers, tokenColorToDeviceColor } from './transform-util.js';
 
-const SIZE_IDENTIFIERS = ['fontSizes', 'sizing', 'borderWidth', 'borderRadius', 'blur', 'spread', 'x', 'y', 'dimension'];
+const SIZE_IDENTIFIERS = ['sizing', 'borderWidth', 'borderRadius', 'blur', 'spread', 'x', 'y', 'dimension'];
 const SPACING_IDENTIFIERS = ['spacing'];
 const FONT_FAMILY_IDENTIFIERS = ['fontFamily'];
 const FONT_SIZE_IDENTIFIERS = ['fontSizes', 'fontSize'];
@@ -105,32 +105,38 @@ export function registerDialtoneTransforms (styleDictionary) {
       return ['color'].includes(token.type);
     },
     transform: function (token) {
-      if (token.value === 'transparent') { return '#00ffffff'; }
-      const str = Color(token.value).toHex8();
-      return '#' + str.slice(6) + str.slice(0, 6);
+      return tokenColorToDeviceColor(token.value, DeviceObjectFormat.ANDROID_XML);
+    },
+  });
+
+  styleDictionary.registerTransform({
+    name: 'dt/android/xml/color/modifiers',
+    type: 'value',
+    transitive: true,
+    filter: colorModifiersFilter,
+    transform: function (token) {
+      return deviceTransformColorModifiers(token, DeviceObjectFormat.ANDROID_XML);
     },
   });
 
   styleDictionary.registerTransform({
     name: 'dt/android/compose/color',
     type: 'value',
-    transitive: true,
     filter: function (token) {
       return ['color'].includes(token.type);
     },
     transform: (token) => {
-      if (token.$extensions?.['studio.tokens']?.modify?.type === 'alpha') {
-        const alpha = parseFloat(token.$extensions['studio.tokens'].modify.value);
-        const alphaHex = Math.round(alpha * 255).toString(16).padStart(2, '0');
-        // ex: token.value = Color(0xff222222)
-        const rgb = token.value.slice(10, 16);
-        return `Color(0x${alphaHex + rgb})`;
-      }
-      if (!Color(token.value).isValid()) {
-        return token.value;
-      }
-      const hex8 = Color(token.value).toHex8();
-      return `Color(0x${hex8.slice(6) + hex8.slice(0, 6)})`;
+      return tokenColorToDeviceColor(token.value, DeviceObjectFormat.ANDROID_COMPOSE);
+    },
+  });
+
+  styleDictionary.registerTransform({
+    name: 'dt/android/compose/color/modifiers',
+    type: 'value',
+    transitive: true,
+    filter: colorModifiersFilter,
+    transform: (token) => {
+      return deviceTransformColorModifiers(token, DeviceObjectFormat.ANDROID_COMPOSE);
     },
   });
 
@@ -165,7 +171,10 @@ export function registerDialtoneTransforms (styleDictionary) {
     type: 'value',
     transitive: true,
     filter: function (token) {
-      return [...SPACING_IDENTIFIERS, ...SIZE_IDENTIFIERS].includes(token.type);
+      return [...SPACING_IDENTIFIERS, ...SIZE_IDENTIFIERS].includes(token.type) &&
+        // The fontSize token in typography tokens is a 'dimension' type for some reason,
+        // so have this special case to exclude it from this transform.
+        !FONT_SIZE_IDENTIFIERS.includes(token.name);
     },
     transform: (token) => {
       const floatVal = parseFloat(token.value);
@@ -184,7 +193,10 @@ export function registerDialtoneTransforms (styleDictionary) {
     type: 'value',
     transitive: true,
     filter: function (token) {
-      return [...FONT_SIZE_IDENTIFIERS].includes(token.type);
+      return [...FONT_SIZE_IDENTIFIERS].includes(token.type) ||
+        // The fontSize token in typography tokens is a 'dimension' type for some reason,
+        // so have this special case to include it in this transform.
+        FONT_SIZE_IDENTIFIERS.includes(token.name);
     },
     transform: (token) => {
       const floatVal = parseFloat(token.value);
@@ -289,11 +301,17 @@ export function registerDialtoneTransforms (styleDictionary) {
       return ['color'].includes(token.type);
     },
     transform: (token) => {
-      const { r, g, b, a } = Color(token.value).toRgb();
-      const rFixed = (r / 255.0).toFixed(3);
-      const gFixed = (g / 255.0).toFixed(3);
-      const bFixed = (b / 255.0).toFixed(3);
-      return `UIColor(red: ${rFixed}, green: ${gFixed}, blue: ${bFixed}, alpha: ${a})`;
+      return tokenColorToDeviceColor(token.value, DeviceObjectFormat.IOS_SWIFT);
+    },
+  });
+
+  styleDictionary.registerTransform({
+    name: 'dt/ios/color/modifiers',
+    type: 'value',
+    transitive: true,
+    filter: colorModifiersFilter,
+    transform: (token) => {
+      return deviceTransformColorModifiers(token, DeviceObjectFormat.IOS_SWIFT);
     },
   });