+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [9.101.2](https://github.com/dialpad/dialtone/compare/dialtone/v9.101.1...dialtone/v9.101.2) (2025-02-19)
+
+
+### Bug Fixes
+
+* **Tokens:** DLT-2347 mobile color token modifiers fix ([#628](https://github.com/dialpad/dialtone/issues/628)) ([9545644](https://github.com/dialpad/dialtone/commit/9545644dc1837061e57370e06c1e695a56681518))
+* **Tokens:** DLT-2348 android font sizes were dp instead of sp ([#631](https://github.com/dialpad/dialtone/issues/631)) ([adda07b](https://github.com/dialpad/dialtone/commit/adda07bdd52126cac3a60c522f54da82034c3484))
+
+## [9.101.1](https://github.com/dialpad/dialtone/compare/dialtone/v9.101.0...dialtone/v9.101.1) (2025-02-19)
+
+
+### Bug Fixes
+
+* **Docs:** NO-JIRA update the figma link in the button docsite page ([#634](https://github.com/dialpad/dialtone/issues/634)) ([77a3a57](https://github.com/dialpad/dialtone/commit/77a3a571bec9089f0c161f0936f35ce44d850664))
+* NO-JIRA custom item layout ([#639](https://github.com/dialpad/dialtone/issues/639)) ([1553250](https://github.com/dialpad/dialtone/commit/155325090d5f8ac38b0b9de377a6a9731a9274ae))
+
+# [9.101.0](https://github.com/dialpad/dialtone/compare/dialtone/v9.100.0...dialtone/v9.101.0) (2025-02-17)
+
+
+### Documentation
+
+* NO-JIRA fix ivr-node ([#630](https://github.com/dialpad/dialtone/issues/630)) ([bca81d8](https://github.com/dialpad/dialtone/commit/bca81d8ac1f0701512a35c6af704762c08d8f080))
+
+
+### Features
+
+* **Design Tokens:** NO-JIRA experimental deca token set updates again ([#632](https://github.com/dialpad/dialtone/issues/632)) ([c7ad513](https://github.com/dialpad/dialtone/commit/c7ad513f49215728dae86bf3305dd54f44df1251))
+
 # [9.100.0](https://github.com/dialpad/dialtone/compare/dialtone/v9.99.0...dialtone/v9.100.0) (2025-02-11)